adjusted file structure, added homepage route
