+++ controllers/api/reviewRoutes.js
@@ -1,5 +1,5 @@
-const router = require('express').Router();
-const { Review } = require('../../models');
-const withAuth = require('../../utils/auth');
+// const router = require('express').Router();
+// const { Review } = require('../../models');
+// const withAuth = require('../../utils/auth');
 
-module.exports = router;
+// module.exports = router;
