+++ views/layouts/main.handlebars
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>It's A Hardback Life Book Club</title>
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.9.3/css/bulma.min.css">
-  <link rel="stylesheet" href="./public/css/style.css">
+  <link rel="stylesheet" href="./css/style.css">
   <script defer src="https://friconix.com/cdn/friconix.js"> </script>
 </head>
